+++ PersonalFinance/PersonalFinance.PublicWeb/App/controllers/budgets/budgetListController.ts
@@ -1,6 +1,6 @@
 ﻿module PersonalFinance.Controllers {
     export class BudgetListController {
-        static $inject = ['$scope', 'messageService', '$validation', 'budgetDataService'];
+        static $inject = ['$scope', 'messageService', '$validation', 'budgetDataService', '$uibModal'];
 
         recentBudget: Models.IRecentBudgetResponse;
         budgets: Models.IBudget[];
@@ -14,7 +14,8 @@
             private $scope: IScope,
             private messageService: Modules.MessageDisplay.IMessageService,
             private $validation,
-            private budgetDataService: Services.IBudgetDataService
+            private budgetDataService: Services.IBudgetDataService,
+            private $uibModal
         ) {
             $scope.vm = this;
 
@@ -57,21 +58,43 @@
             });
         }
 
-        setWaitingDelete(budget) {
-            this.delBudget = budget;
+        deleteBudget(budget) {
+            var modalInstance = this.$uibModal.open({
+                animation: true,
+                templateUrl: '/App/views/modals/confirm-modal.html',
+                controller: 'confirmModalController',
+                resolve: {
+                    modalData: () => {
+                        return {
+                            modalHead: 'Delete Budget',
+                            modalBody: 'Are you sure you want to delete this budget?'
+                        };
+                    }
+                }
+            });
+
+            modalInstance.result.then(confirm => {
+                if (confirm) {
+                    this.budgetDataService.deleteBudget(budget).then(
+                    (response) => {
+                        this.messageService.addInfo(response.data.response);
+                        this.messageService.removeInfoAfterSeconds(response.data.response, 3);
+                        this.loadBudgets();
+                    }, (err) => {
+                        var message = "Failed to delete budget";
+                        this.messageService.addError(message);
+                        this.messageService.removeErrorAfterSeconds(message, 4);
+                    });
+                }
+            });
         }
 
-        deleteBudget() {
-            this.budgetDataService.deleteBudget(this.delBudget).then(
-            (response) => {
-                this.messageService.addInfo(response.data.response);
-                this.messageService.removeInfoAfterSeconds(response.data.response, 3);
-                this.loadBudgets();
-            }, (err) => {
-                var message = "Failed to delete budget";
-                this.messageService.addError(message);
-                this.messageService.removeErrorAfterSeconds(message, 4);
-            });
+        getRemainingForBudget(budget: Models.IBudget) {
+            if (budget.type == 'Savings') {
+                return budget.allocatedAmount - budget.balance;
+            } else {
+                return budget.balance;
+            }
         }
     }
 }
